refact(Flash messages): reorganize redux files
