+++ src/components/customers/list/destroy.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { withRouter, Link } from 'react-router-dom';
 
-import { addFlashMessage } from 'actions/flash-messages.js';
+import { addFlashMessage } from 'services/flash-messages';
 import { deleteCustomer } from 'actions/customers.js';
 import { getCurrentOrganizationId } from 'selectors/organizations.js';
 import { confirm } from 'components/utils/confirm';
