+++ src/components/login/login.jsx
@@ -4,7 +4,7 @@ import { connect } from 'react-redux'
 import { Redirect } from 'react-router-dom';
 import { Panel } from 'react-bootstrap'
 
-import { addFlashMessage } from 'actions/flash-messages.js';
+import { addFlashMessage } from 'services/flash-messages';
 import { loginUser } from 'services/auth';
 import { prepareSubmissionError } from 'utils/errors';
