+++ src/components/organizations/list/list-item.jsx
@@ -4,7 +4,7 @@ import { connect } from 'react-redux';
 import { withRouter } from 'react-router-dom';
 import { ListGroupItem } from 'react-bootstrap';
 
-import { addFlashMessage } from 'actions/flash-messages.js';
+import { addFlashMessage } from 'services/flash-messages';
 import { setCurrentOrganization } from 'actions/organizations.js';
 
 class OrganizationsListItem extends React.Component {
