+++ src/components/user/cancel-account.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 
-import { addFlashMessage } from 'actions/flash-messages.js';
+import { addFlashMessage } from 'services/flash-messages';
 import { cancelAccount as cancelAccountAction } from 'actions/users.js';
 
 class CancelAccount extends React.Component {
