+++ src/reducers/index.ts
@@ -1,6 +1,6 @@
 export { default as app }                  from './app';
 export { default as auth }                 from 'services/auth/reducer.js';
-export { default as flashMessages }        from './flash-messages';
+export { default as flashMessages }        from 'services/flash-messages/reducer.js';
 export { default as invoice }              from './invoice';
 export { default as invoices }             from './invoices';
 export { default as organizations }        from './organizations';
